# manufacturing-ai-cost-platform/docs/API_CONTRACT.yaml
openapi: 3.0.3
info:
  title: Manufacturing AI Cost Intelligence API
  version: 1.2.0
  description: >
    API contract for the Manufacturing AI Cost Intelligence & Autonomous
    Optimization Platform.

    Field types and persistence semantics follow DATABASE_SCHEMA.md. Cost values
    are always labelled with their provenance (ACTUAL, ESTIMATED, FORECAST,
    SIMULATED, UNAVAILABLE) so they are never presented interchangeably.
servers:
  - url: /api/v1
...
paths:
...
  /cost/summary:
    get:
      tags: [Costs]
      summary: Get current cost summary
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PlantId"
        - $ref: "#/components/parameters/DepartmentId"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Cost summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/InternalError"
  /cost/by-model:
    get:
      tags: [Costs]
      summary: Get cost grouped by model
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PlantId"
        - $ref: "#/components/parameters/DepartmentId"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Model cost breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostBreakdown"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/InternalError"
  /cost/by-agent:
    get:
      tags: [Costs]
      summary: Get cost grouped by agent
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PlantId"
        - $ref: "#/components/parameters/DepartmentId"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Agent cost breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostBreakdown"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/InternalError"
  /cost/by-plant:
    get:
      tags: [Costs]
      summary: Get cost grouped by plant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Plant cost breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostBreakdown"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/InternalError"
  /cost/trend:
    get:
      tags: [Costs]
      summary: Get historical cost trend
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PlantId"
        - $ref: "#/components/parameters/DepartmentId"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
      responses:
        "200":
          description: Historical cost trend
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostTrend"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PlantId:
      name: plant_id
      in: query
      schema:
        type: string
    DepartmentId:
      name: department_id
      in: query
      schema:
        type: string
    From:
      name: from
      in: query
      schema:
        type: string
        format: date-time
    To:
      name: to
      in: query
      schema:
        type: string
        format: date-time
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: >
            Safe human-readable summary. Never contains stack traces, internal
            exception detail or secrets.
        request_id:
          type: string
        details:
          type: object
          description: Field-level validation detail where applicable.
...
    CostSummary:
      type: object
      description: >
        Actual and estimated spend are reported separately and never summed into
        a single unlabelled figure (AI_DEVELOPMENT_RULES.md sections 41 and 42).
      properties:
        actual_cost:
          type: number
          description: Sum of cost_events with provenance = ACTUAL.
        estimated_cost:
          type: number
          description: Sum of cost_events with provenance = ESTIMATED.
        unavailable_cost_events:
          type: integer
          description: Count of events with provenance = UNAVAILABLE.
        currency:
          type: string
          description: Platform base currency used for aggregation.
        total_requests:
          type: integer
        total_tokens:
          type: integer
        average_cost_per_request:
          type: number
        budget_consumed_percent:
          type: number
        forecast_month_end_cost:
          type: number
          description: FORECAST value. Must be displayed as a forecast.
    CostBreakdownItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        actual_cost:
          type: number
        estimated_cost:
          type: number
        currency:
          type: string
        total_requests:
          type: integer
        total_tokens:
          type: integer
    CostBreakdown:
      type: object
      properties:
        dimension:
          type: string
          enum: [model, agent, plant]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CostBreakdownItem"
    CostTrendPoint:
      type: object
      properties:
        bucket_start:
          type: string
          format: date-time
        actual_cost:
          type: number
        estimated_cost:
          type: number
        currency:
          type: string
        total_requests:
          type: integer
        total_tokens:
          type: integer
    CostTrend:
      type: object
      properties:
        granularity:
          type: string
          enum: [hour, day, week, month]
        points:
          type: array
          items:
            $ref: "#/components/schemas/CostTrendPoint"
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authorization failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"